Removed space so that matching of merknader works properly
There was an extra space in the merknad "gjenutførsel  eller retur", so this PR fixed that.

We are using this when mapping the legal merknad to be used, so it cannot have typos. It would probably be nice to have with a test that handles this if there were new merknader added later.